chore(storage node): move fakeclient test of storage node to controllers package
